fix: refactor useList
